requirements.txt
Django==2.2.2
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-js-asset==1.2.2
django-tempus-dominus==5.1.2.5
gunicorn==19.9.0
Pillow==8.0.1
pytz==2019.1
SQLAlchemy==1.3.4
sqlparse==0.3.0
whitenoise==4.1.2
